Fix various bugs and refactor
